// src/app/shared/generated/model/createWorkspaceRoleRequest.ts
/**
 * onecx-workspace-bff
 * OneCX Workspace BFF
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface CreateWorkspaceRoleRequest { 
    workspaceId: string;
    name?: string;
    description?: string;
}